refactor: 입력 모듈 분리 -입력 받기, 처리, 검증
