+++ src/index.js
@@ -1,78 +1,27 @@
 import RacingCarGame from "./game.js";
-import {
-  validateInputsLength,
-  validateInputsUnique,
-  validateNaturalNumber,
-} from "./common/validate.js";
-import { trimStringSpace, convertStringToArray, convertStringToNumber } from "./common/util.js";
-import { alertMessage } from "./common/message.js";
-const { INVALID_LENGTH_MESSAGE, DUPLICATE_NAME_MESSAGE, INVALID_NUMBER_MESSAGE } = alertMessage;
+import InputHandler from "./handler/inputHandler.js";
+import { convertStringToNumber } from "./common/util.js";
 export default class App {
   constructor() {
     this.carNames = [];
     this.racingCount = 0;
 
-    this.carNameInputField = document.getElementById("car-names-input");
-    this.carNameSubmitButton = document.getElementById("car-names-submit");
+    this.inputHandler = new InputHandler();
 
-    this.racingCountInputField = document.getElementById("racing-count-input");
+    this.carNameSubmitButton = document.getElementById("car-names-submit");
     this.racingCountSubmitButton = document.getElementById("racing-count-submit");
     this.racingCountContainer = document.getElementById("racing-count-container");
-
     this.gameResultContainer = document.getElementById("game-result-container");
-  }
-
-  getCarNamesInput() {
-    const carNamesInput = this.carNameInputField.value;
-    return carNamesInput;
-  }
-
-  parseCarNames(carNamesInput) {
-    const carNamesString = trimStringSpace(carNamesInput); // 공백제거
-    const carNames = convertStringToArray(carNamesString);
-    return carNames;
-  }
-
-  validateCarNames(carNames) {
-    const isValidLength = validateInputsLength(carNames);
-    const isUnique = validateInputsUnique(carNames);
-    const isValid = isValidLength && isUnique;
-
-    if (!isValidLength) {
-      alert(INVALID_LENGTH_MESSAGE);
-      return;
-    }
-
-    if (!isUnique) {
-      alert(DUPLICATE_NAME_MESSAGE);
-    }
-
-    return isValid;
   }
 
   displayRacingCountContainer() {
     this.racingCountContainer.style.display = "block";
   }
 
-  getRacingCountInput() {
-    const racingCountInput = this.racingCountInputField.value;
-    return racingCountInput;
-  }
-
-  validateRacingCount(number) {
-    const isValidNumber = validateNaturalNumber(number);
-
-    if (!isValidNumber) {
-      alert(INVALID_NUMBER_MESSAGE);
-    }
-
-    return isValidNumber;
-  }
-
   submitCarNames() {
-    const carNamesInput = this.getCarNamesInput();
-    const carNames = this.parseCarNames(carNamesInput);
-    const isValidCarNames = this.validateCarNames(carNames);
+    const carNamesInput = this.inputHandler.getCarNamesInput();
+    const carNames = this.inputHandler.parseCarNames(carNamesInput);
+    const isValidCarNames = this.inputHandler.validateCarNames(carNames);
 
     if (!isValidCarNames) {
       return;
@@ -83,9 +32,9 @@ export default class App {
   }
 
   submitRacingCount() {
-    const racingCountInput = this.getRacingCountInput();
+    const racingCountInput = this.inputHandler.getRacingCountInput();
     const racingCount = convertStringToNumber(racingCountInput);
-    const isValidRacingCount = this.validateRacingCount(racingCount);
+    const isValidRacingCount = this.inputHandler.validateRacingCount(racingCount);
 
     if (!isValidRacingCount) {
       return;
